Make ADAMVCFOutputFormat public
This commit makes the ADAMVCFOutputFormat defined in
ADAMVariationContext public and moves it to its own source file.
The alternate definition of ADAMVCFOutputFormat (unused) is
deleted.
